Island: Move IAgentLogRepository into log_service
