+++ monkey/monkey_island/cc/repositories/__init__.py
@@ -9,7 +9,6 @@ from .i_machine_repository import IMachineRepository
 from .i_agent_repository import IAgentRepository
 from .i_node_repository import INodeRepository
 from .i_agent_event_repository import IAgentEventRepository
-from .i_agent_log_repository import IAgentLogRepository
 from .i_agent_plugin_repository import IAgentPluginRepository
