+++ monkey/monkey_island/cc/services/log_service/__init__.py
@@ -1,0 +1,1 @@
+from .i_agent_log_repository import IAgentLogRepository
