+++ monkey/monkey_island/cc/services/log_service/i_agent_log_repository.py
