+++ monkey/tests/monkey_island/in_memory_agent_log_repository.py
@@ -1,5 +1,6 @@
 from common.types import AgentID
-from monkey_island.cc.repositories import IAgentLogRepository, UnknownRecordError
+from monkey_island.cc.repositories import UnknownRecordError
+from monkey_island.cc.services.log_service import IAgentLogRepository
 
 
 class InMemoryAgentLogRepository(IAgentLogRepository):
