+++ monkey/tests/unit_tests/monkey_island/cc/repositories/test_file_agent_log_repository.py
@@ -7,11 +7,11 @@ from tests.monkey_island import InMemoryFileRepository, OpenErrorFileRepository
 
 from monkey_island.cc.repositories import (
     FileAgentLogRepository,
-    IAgentLogRepository,
     IFileRepository,
     RetrievalError,
     UnknownRecordError,
 )
+from monkey_island.cc.services.log_service import IAgentLogRepository
 
 LOG_CONTENTS = "lots of useful information"
 AGENT_ID_1 = UUID("6bfd8b64-43d8-4449-8c70-d898aca74ad8")
